+++ scim2-sdk-client/src/main/java/com/unboundid/scim2/client/requests/ModifyRequestBuilder.java
@@ -72,7 +72,7 @@ public abstract class ModifyRequestBuilder<T extends ModifyRequestBuilder<T>>
    */
   @Override
   @NotNull
-  Invocation.Builder buildRequest()
+  protected Invocation.Builder buildRequest()
   {
     Invocation.Builder request = super.buildRequest();
     if (version != null)